Preparing for open source, first pass (#89)
* remove sbt scaffolding
* consolidate code
* cleanup code, eliminate closed-source sbt image
* remove jfrog references
* proto generate once
* exclude ChiefOfStateService from cov
* remove kafka broker setting
* remove namely from readme

+++ code/service/src/main/scala/com/namely/chiefofstate/Util.scala
@@ -28,7 +28,8 @@ object Util {
   def toCosMetaData(metaData: MetaData): CosMetaData = {
     CosMetaData(
       entityId = metaData.entityId,
-      revisionNumber = metaData.revisionNumber,
+      // TODO: remove .toInt
+      revisionNumber = metaData.revisionNumber.toInt,
       revisionDate = metaData.revisionDate,
       data = metaData.data
     )
